Rois Pasting

+++ ol3-viewer/src/ome/ol3/utils/Conversion.js
@@ -919,8 +919,8 @@ ome.ol3.utils.Conversion.convertPointStringIntoCoords = function(points) {
         if (tok === '') continue;
         var c = tok.split(",");
         if (c.length < 2) return null;
-        var x = parseInt(c[0]);
-        var y = parseInt(c[1]);
+        var x = parseFloat(c[0]);
+        var y = parseFloat(c[1]);
         // NaNs are not acceptable
         if (isNaN(x) || isNaN(y)) return null;
         ret.push([x, -y]);
